validado

+++ app/Http/Controllers/DireccionController.php
@@ -9,27 +9,34 @@ use Illuminate\Http\Request;
 
 class DireccionController extends Controller
 {
-    public function store(Request $request)
-    {
-        $cliente_id = $request->cliente_id; // Ahora obtenemos cliente_id directamente del Request
+        public function store(Request $request)
+        {
+            $cliente_id = $request->cliente_id; // Ahora obtenemos cliente_id directamente del Request
 
-        $request->validate([
-            'cliente_id' => 'required|exists:clientes,id', // Asegúrate de validar cliente_id también
-            'tipo' => 'required|max:255',
-            'direccion' => 'required|max:255',
-            'ubicacion_id' => 'required|integer|exists:ubicaciones,id',
-        ]);
+            $request->validate([
+                'cliente_id' => 'required|exists:clientes,id', // Asegúrate de validar cliente_id también
+                'tipo' => 'required|max:255',
+                'direccion' => 'required|max:255',
+                'ubicacion_id' => 'required|integer|exists:ubicaciones,id',
+            ]);
 
-        $direccion = new Direccion([
-            'cliente_id' => $cliente_id,
-            'tipo' => $request->tipo,
-            'direccion' => $request->direccion,
-            'ubicacion_id' => $request->ubicacion_id,
-        ]);
-        $direccion->save();
+            $direccion = new Direccion([
+                'cliente_id' => $cliente_id,
+                'tipo' => $request->tipo,
+                'direccion' => $request->direccion,
+                'ubicacion_id' => $request->ubicacion_id,
+            ]);
+            $direccion->save();
 
-        return redirect()->route('clientes.show', $cliente_id)->with('success', 'Dirección agregada con éxito.');
-    }
+            // Suponiendo que tienes un modelo Cliente que está relacionado con Direcciones
+            $cliente = Cliente::with('direcciones')->find($cliente_id);
+
+            return response()->json([
+                'success' => true,
+                'message' => 'Dirección agregada con éxito.',
+                'direcciones' => $cliente->direcciones, // Enviamos las direcciones del cliente
+            ]);
+        }
         public function edit($id)
         {
             $direccion = Direccion::findOrFail($id);
@@ -59,6 +66,13 @@ class DireccionController extends Controller
         public function destroy($id)
         {
             $direccion = Direccion::findOrFail($id);
+
+            // Verificar si la dirección está siendo usada en pedidos
+            if ($direccion->pedidos()->count() > 0) {
+                // No permitir la eliminación y redirigir con un mensaje
+                return redirect()->back()->with('error', 'Esta dirección no se puede eliminar porque está siendo usada en pedidos.');
+            }
+
             $clienteId = $direccion->cliente_id;
             $direccion->delete();
